+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 Version 1.1-SNAPSHOT
 -------------
 
+ADDED: 
+
+- new TypeAdapter class BigDecimalMoneyScaledAdapter for BigDecimal that is indented for money values created
+
+CHANGED:
+
+- removed dependency org.json:json
+
 Version 1
 -------------